FLIZ-83 docs: 변경 구조에 대한 스토리북 재작성
